requirements.txt
# OFB Intelligent Document Discovery System v2.0
# Production dependencies

# ── Web Framework ────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── Background Scheduler ────────────────────────────────────────
apscheduler>=3.10.0

# ── Vector Store ─────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── DB Connectivity ──────────────────────────────────────────────
mysql-connector-python>=8.3.0

# ── Numerics ─────────────────────────────────────────────────────
numpy>=1.24.0

# ── LangChain / LangGraph ───────────────────────────────────────
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.3.0

# ── Reranking ────────────────────────────────────────────────────
flashrank>=0.2.0

# ── Authentication & Encryption ──────────────────────────────────
bcrypt>=4.1.0
PyJWT>=2.8.0
cryptography>=42.0.0

# ── File Text Extraction ─────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
docx2txt>=0.8

# ═══════════════════════════════════════════════════════════════════
# Ollama models (pull ONCE, then fully offline):
#   ollama pull nomic-embed-text   # ~274 MB  — embeddings
#   ollama pull llama3.1:8b        # ~4.7 GB  — RAG + schema intelligence
# ═══════════════════════════════════════════════════════════════════
